refactor: 전역에서 관리되는 navigation store는 app 디렉토리 안에
